+++ core/CHANGELOG.md
@@ -7,6 +7,7 @@
 * Passing networks by id instead of name inside docker manager
 * Added a "Debugging failed tests" tutorial
 * Bugfix for broken CI checks that don't verify CHANGELOG is actually modified
+* Pass network ID instead of network name to the controller
 
 # 0.7.0
 * Allow developers to configure how wide their test networks will be